[Model] Reorganize pooling layers
